% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scenario.R
\name{get_scenario}
\alias{get_scenario}
\title{Get scenario name from concatenated scenario string}
\usage{
get_scenario(scenario_parameter)
}
\arguments{
\item{scenario_parameter}{A string containing the scenario source and name,
separated by a single underscore}
}
\value{
the scenario name part of the string
}
\description{
Get the Scenario name from a concatenated scenario string (for example,
"NZE_2050" from "WEO2022_NZE_2050")
}
