% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scenario.R
\name{get_scenario_source}
\alias{get_scenario_source}
\title{Get scenario source from concatenated scenario string}
\usage{
get_scenario_source(scenario_parameter)
}
\arguments{
\item{scenario_parameter}{A string containing the scenario source and name,
separated by a single underscore}
}
\value{
the scenario source part of the string
}
\description{
Get the Scenario source from a concatenated scenario string (for example,
"WEO2022" from "WEO2022_NZE_2050")
}
