% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacta.interactive.report-package.R,
%   R/pacta.portfolio.report-package.R
\docType{package}
\name{pacta.json.export-package}
\alias{pacta.json.export-package}
\title{pacta.json.export: Export PACTA results as JSON files}
\description{
Convert PACTA analysis results to JSON files suitible for use in interactive plots.

Convert PACTA analysis results to JSON files suitible for use in interactive plots.
}
\author{
\strong{Maintainer}: Alex Axthelm \email{aaxthelm@rmi.org} (\href{https://orcid.org/0000-0001-8579-8565}{ORCID}) [contractor]

Authors:
\itemize{
  \item Monika Furdyna \email{monika.furdyna@gmail.com} (\href{https://orcid.org/0000-0002-3728-0646}{ORCID}) [contractor]
  \item CJ Yetman \email{cj@cjyetman.com} (\href{https://orcid.org/0000-0001-5099-9500}{ORCID}) [contractor]
  \item Jackson Hoffart \email{jackson.hoffart@gmail.com} (\href{https://orcid.org/0000-0002-8600-5042}{ORCID}) [contractor]
  \item Jacob Kastl \email{jacob.kastl@gmail.com} [contractor]
}

Other contributors:
\itemize{
  \item Rocky Mountain Institute \email{PACTA4investors@rmi.org} [copyright holder, funder]
}

}
\keyword{internal}
