% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios_found_in_sectors.R
\name{scenarios_found_in_sectors}
\alias{scenarios_found_in_sectors}
\title{Generate filtering vector for scenarios in sectors}
\usage{
scenarios_found_in_sectors(data, select_scenario_param, sectors)
}
\arguments{
\item{data}{data.frame with columns "ald_sector", "scenario", and
"scenario_source"}

\item{select_scenario_param}{Concatenated Scenario string (source and name,
ex. "WEO2022_NZE_2050"). Rows in \code{data} matching this scenario source and
name will be included in the output.}

\item{sectors}{A character vector of PACTA sector names. Rows in \code{data}
matching any of these sectors will be included in the output.}
}
\value{
Logical vector, of length equal to the number of rows in data,
indicating whether each row is in the specified sectors and scenarios.
Suitible for use as a filtering function in \code{dplyr::filter()}.
}
\description{
For a portfolio-shaped data.frame (with columns "ald_sector", "scenario",
and "scenario_source"), generate a logical vector indicating whether each
row is in the specified sectors and scenarios.
}
